pyproject.toml
[tool.poetry]
name = "sentinela-core"
version = "0.1.0"
description = "Sentinela-Core: EDR open-source con IA local para laboratorios de hacking ético."
authors = ["Benjamín Palma <be.palmap@duocuc.cl>"]
readme = "README.md"
packages = [{ include = "sentinela_core" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

pydantic-settings = "^2.0.0"
sqlalchemy = "^2.0.0"
pymysql = "^1.1.0"
psutil = "^5.9.0"
yara-python = "^4.5.0"
scapy = "^2.5.0"
typer = "^0.12.0"
python-dotenv = "^1.0.0"
fastapi = "^0.124.2"
uvicorn = "^0.38.0"
requests = "^2.32.5"
jinja2 = "^3.1.6"
python-multipart = "^0.0.20"
aiofiles = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
sentinela = "sentinela_core.cli:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
